requirements.txt
beautifulsoup4==4.12.2
boto3==1.28.56
botocore==1.31.56
bs4==0.0.1
greenlet==2.0.2
jmespath==1.0.1
playwright==1.38.0
psycopg2==2.9.8
pyee==9.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
soupsieve==2.5
typing_extensions==4.8.0
urllib3==1.26.16
